fix bug and data root

+++ tools/test.py
@@ -135,7 +135,7 @@ def main():
              sv_dir=final_output_dir)
 
     end = timeit.default_timer()
-    logger.info('Mins: %d' % np.int((end-start)/60))
+    logger.info('Mins: %d' % np.int_((end-start)/60))
     logger.info('Done')
